Revert "Feature/loginreg"
